Normative: Add yearOfWeek calendar methods and properties
The rationale that we got from feedback during implementation is that it's
not obvious that the weekOfYear may belong to a different "week calendar
year" than the date's calendar year, which may lead to bugs like
unwittingly presenting the date 2022-01-01 as being in week 52 of 2022
(it is actually in week 52 of 2021.)

Additionally, the yearOfWeek isn't that straightforward to calculate
unless you are already calculating the week number.

Adds the following APIs:
  Temporal.PlainDate.yearOfWeek
  Temporal.PlainDateTime.yearOfWeek
  Temporal.ZonedDateTime.yearOfWeek
  Temporal.Calendar.yearOfWeek()
and updates the ToISOWeekOfYear and CalendarDateWeekOfYear abstract ops to
return a Record with [[Week]] and [[Year]] fields, instead of just a week
number.

Closes: #2405